requirements.txt
pandas
plotly
kaleido

streamlit_js_eval
